Time GPU code using CUDA events

+++ Distributed-Computing/scan.cuh
@@ -1,5 +1,5 @@
 long blockscan(int *output, int *input, int length, bool bcao);
-long scan(int *output, int *input, int length, bool bcao);
+float scan(int *output, int *input, int length, bool bcao);
 
 void scanLargeDeviceArray(int *output, int *input, int length, bool bcao);
 void scanSmallDeviceArray(int *d_out, int *d_in, int length, bool bcao);
